+++ src/beatmap/schema/v4/declaration.ts
@@ -380,6 +380,8 @@ export const ArcSchema: {
 
 /**
  * Schema declaration for v4 `Spawn Rotation`.
+ *
+ * @deprecated removed as of 1.39, convert to `r` in object lane
  */
 export const SpawnRotationSchema: {
    readonly [key in keyof ISpawnRotation]: ISchemaDeclaration;